+++ internal/builtins/builtins.go
@@ -52,5 +52,6 @@ func SurveyorRegistry() *surveyor.Registry {
 	reg.Register(dsurveyors.NewK8sCluster())
 	reg.Register(dsurveyors.NewGitHubOrgRepos())
 	reg.Register(dsurveyors.NewGitLabGroupProjects())
+	reg.Register(dsurveyors.NewAzureDevOpsRepos())
 	return reg
 }